apps/scraper/requirements.txt
supabase==2.10.0
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
